+++ server/account/User.cs
@@ -1,3 +1,5 @@
+#nullable disable
+
 using System;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Diagnostics.CodeAnalysis;
@@ -10,14 +12,14 @@ namespace SecurityCameraServer;
 public class User
 {
     public int ID { get; set; }
-    public string? EMail { get; set; }
-    public string? Name { get; set; }
-    public string? PasswordHash { get; private set; }
+    public string EMail { get; set; }
+    public string Name { get; set; }
+    public string PasswordHash { get; private set; }
 
-    public List<Camera>? Cameras { get; set; }
+    public List<Camera> Cameras { get; set; }
 
     [NotMapped]
-    public SecureHash<SHA256>? PasswordHashData
+    public SecureHash<SHA256> PasswordHashData
     {
         get => SecureHash<SHA256>.FromHexString(PasswordHash);
         set => PasswordHash = value?.ToString();