+++ server/camera/Camera.cs
@@ -1,4 +1,4 @@
-
+#nullable disable
 
 using Microsoft.EntityFrameworkCore;
 
@@ -7,8 +7,8 @@ namespace SecurityCameraServer
 	public class Camera
 	{
 		public int ID { get; set; }
-		public string Name { get; set; }
-		public string CameraGUID { get; set; }
-		public User Owner { get; set; }
+		public string Name { get; init; }
+		public string CameraGuid { get; init; }
+		public User Owner { get; init; }
 	}
 }