+++ server/camera/RegisterCameraData.cs
@@ -1,9 +1,11 @@
+#nullable disable
+
 namespace SecurityCameraServer;
 
 public class RegisterCameraData
 {
 	public string Name { get; set; }
-	public string? RequestedGUID { get; set; }
+	public string RequestedGUID { get; set; }
 	
 	// other data like framerate??
 }