+++ server/session/CameraSession.cs
@@ -4,21 +4,20 @@ namespace SecurityCameraServer;
 
 public class CameraSession
 {
-	private static HttpClient _client = new HttpClient();
-	public Camera Camera { get; set; }
-	public WebSocket? Socket { get; set; }
-	public CancellationTokenSource? CancellationTokenSource { get; set; }
+	public CameraSession(Camera camera, WebSocket socket)
+	{
+		Camera = camera;
+		Socket = socket;
+	}
+	
+	public Camera Camera { get; init; }
+	public WebSocket Socket { get; init; }
+	public CancellationTokenSource? CancellationTokenSource { get; init; }
 	public byte[]? CurrentSnapshot { get; set; }
 	public DateTime LastSnapshotReceived { get; set; }
 	
-	public Action<CameraSession>? OnSnapshotReceived;
+	public Action<CameraSession>? OnSnapshotReceived { get; init; }
 
-	// public async Task<byte[]?> RequestSnapshotAsync()
-	// {
-	// 	var message = new HttpRequestMessage(HttpMethod.Get, $"{CameraURI}/snapshot");
-	// 	var response = await _client.SendAsync(message);
-	// 	
-	// 	return await response.Content.ReadAsByteArrayAsync();
-	// }
+	public List<CameraViewerSession> Subscribed { get; } = new();
 
 }